pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = [ "py36", "py37", "py38",]
exclude = "((.eggs | .git | .pytest_cache | build | dist))"

[tool.poetry]
name = "tradingsignal"
version = "0.0.0"
description = "Open source framework to automate crypto currency trading & technical analysis: heikin ashi, RSI, MACD, MA, PA, connect to telegram, slack and more"
authors = ["Trading Signal <Trading.SignalX@gmail.com>"]
repository = "https://github.com/TradingSignal/TradingSignal"
readme = "README.md"
license = "Apache-2.0"

[tool.poetry.dependencies]
python = ">=3.6.1,<3.9"
"ruamel.yaml" = "^0.16.12"
APScheduler = "^3.6.3"
pandas = "^1.0.4"
backports-datetime-fromisoformat = "^1.0.0"
oandapyV20 = "^0.6.3"
TA-Lib = "^0.4.19"
requests = "^2.25.0"
fastapi = "^0.61.2"
uvicorn = "^0.12.2"

[tool.poetry.dev-dependencies]
pytest-cov = "^2.10.0"
pytest-xdist = "^1.32.0"
pytest = "^5.3.4"
black = "^20.8b1"
flake8 = "^3.8.4"

[tool.poetry.scripts]
tradingsignal = "tradingsignal.__main__:main"